pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aigate"
version = "0.1.0"
description = "AI Prompt Secret Scanner — local proxy that blocks secrets before they reach AI APIs"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "mitmproxy>=10.0",
]

[project.scripts]
aigate = "aigate.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
